"Top-accent" anchors; Latin-1 accent spacing characters
1. Moved "top-accent" anchor in h and k to a point directly above the
stem, at the same coordinates as the "top-accent" anchor for l.

2. Modified macron and cedilla glyphs in Latin-1 Supplement so that
they are "spacing characters" (which they should be according to the
Unicode specification) rather than "modifying characters". I made the
macron have the same bearings as the dieresis, and the cedilla have the
same bearings as the cedilla in ç.