Up to uni0255
Added anchor points to script a, adjusted right bearing.

Adjusted turned script a, remodelling the bowl on those of thorn and b,
and the serifs on those of p and b. Added anchor points.

Adjusted bearings of implosive b, added anchor points.

Deleted superfluous anchor points on open o, remodelled the upper serif
on that of reversed-s, and increased left bearing.

Adjusted thickness of tail on c-with-tail, changed bearings, added
anchor points.