tests: Add a (multilingual) test for some (almost) real world Mathematics.
The section on projective planes I contributed to Wikipedia a long time ago,
and I think that it is a useful "real world" test.

The section on field extensions is written in Portuguese, which is a
language that I (sometimes) use, to get a feel for the blending of how the
mathematics parts go with the non-mathematical parts.

The ideal case would be to get rid of Asana Math as much as possible.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>